chore: remove and clean components
