+++ README.md
@@ -1,1 +1,34 @@
-# ToDo List with react context
+# TodosList with react-context
+
+## Live version
+
+[TodosList on netlify](https://gorgeous-gingersnap-a97472.netlify.app/)
+
+## Contents
+
+In this TodosList project we have the following sections done:
+
+- [x] Todo
+- [x] Todo List
+
+## Technologies Used
+
+This project is mainly based on [React.js](https://reactjs.org/) plus others few libraries listed below:
+
+- [Tailwind css](https://tailwindcss.com/)
+
+## How To Run The Project
+
+First, install the project's dependencies:
+
+```
+npm install
+```
+
+Then, run the development server:
+
+```
+npm start
+```
+
+There you go 😊, You can preview the project on [http://localhost:3000](http://localhost:3000)
